possible fixes

+++ assignment 2/problem3.cpp
@@ -1,6 +1,6 @@
 #include <bits/stdc++.h>
 
-long long *insert(const long long *arr, int &currentSize, long long value) {
+long long *insert(const long long *arr, long long &currentSize, long long value) {
     long long *newArr = new long long[currentSize + 1];
     for (int i = 0; i < currentSize; i++) {
         newArr[i] = arr[i];
@@ -11,7 +11,7 @@ long long *insert(const long long *arr, int &currentSize, long long value) {
     return newArr;
 }
 
-long long *pop(long long *arr, int pos, int &currentSize) {
+long long *pop(long long *arr, long long pos, long long &currentSize) {
     if (pos > currentSize || (pos == 0 && currentSize == 0)) {
         return arr;
     }
@@ -30,7 +30,7 @@ long long *pop(long long *arr, int pos, int &currentSize) {
 
 int main() {
     long long *arr = nullptr;
-    int currentSize = 0;
+    long long currentSize = 0;
 
     std::string s;
     while (std::getline(std::cin, s)) {
@@ -40,7 +40,7 @@ int main() {
             long long value = std::stoll(s.substr(7, s.length()));
             arr = insert(arr, currentSize, value);
         } else if (s.find("pop") == 0) {
-            int pos = std::stoi(s.substr(4, s.length()));
+            long long pos = std::stoi(s.substr(4, s.length()));
             arr = pop(arr, pos, currentSize);
         } else if (s.find("print") == 0) {
             for (int i = 0; i < currentSize; i++) {
@@ -49,7 +49,7 @@ int main() {
             std::cout << std::endl;
         } else if (s.find("swap") == 0) {
             std::string trash;
-            int pos1, pos2;
+            long long pos1, pos2;
             std::stringstream ss(s);
             ss >> trash >> pos1 >> pos2;
             if (pos1 > currentSize || pos2 > currentSize || (pos1 == 0 && pos2 == 0 && currentSize == 0))
